исправил плоское решение

+++ scripts.py
@@ -16,11 +16,15 @@ def find_schoolkid_by_name(schoolkid_name):
 
 def fix_marks_on_five(schoolkid_name):
     schoolkid = find_schoolkid_by_name(schoolkid_name)
+    if not schoolkid:
+        return None
     Mark.objects.filter(schoolkid=schoolkid, points__in=[2, 3]).update(points=5)
 
 
 def fix_marks(schoolkid_name):
     schoolkid = find_schoolkid_by_name(schoolkid_name)
+    if not schoolkid:
+        return None
     bad_marks = Mark.objects.filter(schoolkid=schoolkid, points__in=[2, 3])
     for bad_mark in bad_marks:
         bad_mark.points = random.choice([4, 5])
@@ -29,12 +33,16 @@ def fix_marks(schoolkid_name):
 
 def remove_chastisements(schoolkid_name):
     schoolkid = find_schoolkid_by_name(schoolkid_name)
+    if not schoolkid:
+        return None
     bad_opinions = Chastisement.objects.filter(schoolkid=schoolkid)
     bad_opinions.delete()
 
 
 def create_commendation(schoolkid_name, subject_name):
     schoolkid = find_schoolkid_by_name(schoolkid_name)
+    if not schoolkid:
+        return None
     kid_lessons_on_subject = Lesson.objects.filter(year_of_study=schoolkid.year_of_study,
                                                    group_letter=schoolkid.group_letter,
                                                    subject__title=subject_name)
